# Encode decode performance improvements