--- a/README.md
+++ b/README.md
@@ -31,4 +31,9 @@ To run tests, call:
 
 To run benchmarks, call:
 
-    $ make benchmark
+    $ make bench
+
+## Latest benchmarks
+
+    BenchmarkEncode-4        5000000           387 ns/op         109 B/op          5 allocs/op
+    BenchmarkDecode-4        5000000           302 ns/op          19 B/op          2 allocs/op