This improves a lot the performance of `Encode` and `Decode` functions, original benchmarks on my computer were:
```
BenchmarkEncode-4   	 1000000	      1270 ns/op	     120 B/op	      12 allocs/op
BenchmarkDecode-4   	  300000	      4363 ns/op	     314 B/op	      45 allocs/op
```
Now they are:
```
BenchmarkEncode-4   	 5000000	       341 ns/op	     109 B/op	       5 allocs/op
BenchmarkDecode-4   	 5000000	       310 ns/op	      19 B/op	       2 allocs/op
```
Which is a 3x improvement in `Encode` and 13x in `Decode`

Additionally, benchmarks were split into different levels, and I added more points to be encoded/decoded (avoiding any processor-level caching that could be applied).
With those benchmarks, original version benchmarked:
```
BenchmarkEncodeLevel2-4    	 2000000	       761 ns/op	      92 B/op	       7 allocs/op
BenchmarkEncodeLevel6-4    	 1000000	      1258 ns/op	     104 B/op	      11 allocs/op
BenchmarkEncodeLevel15-4   	  500000	      2490 ns/op	     159 B/op	      20 allocs/op
BenchmarkDecodeLevel2-4    	 1000000	      2279 ns/op	     164 B/op	      23 allocs/op
BenchmarkDecodeLevel6-4    	  500000	      3791 ns/op	     269 B/op	      39 allocs/op
BenchmarkDecodeLevel15-4   	  200000	      7365 ns/op	     507 B/op	      75 allocs/op
```
It's clear that time growth is almost linear with the level being encoded/decoded. With this changes it's almost constant (most of time time is wasted on calculating Mercator projection & point coordinates which is constant, then obviously encoding/decoding is linear but it's just a small fraction):
```
BenchmarkEncodeLevel2-4    	 5000000	       328 ns/op	      88 B/op	       5 allocs/op
BenchmarkEncodeLevel6-4    	 5000000	       356 ns/op	      96 B/op	       5 allocs/op
BenchmarkEncodeLevel15-4   	 3000000	       431 ns/op	     144 B/op	       5 allocs/op
BenchmarkDecodeLevel2-4    	 5000000	       295 ns/op	      19 B/op	       2 allocs/op
BenchmarkDecodeLevel6-4    	 5000000	       311 ns/op	      19 B/op	       2 allocs/op
BenchmarkDecodeLevel15-4   	 5000000	       355 ns/op	      19 B/op	       2 allocs/op
```